refactor: Migrate ToggleState to dedicated context and provider
