+++ src/features/post/edit-post/ui/PostEditModal.tsx
@@ -2,7 +2,7 @@ import { Modal } from "../../../../shared/ui"
 import { Post } from "../../../../entities/post/model"
 import { PostForm } from "../../../../entities/post/ui"
 import { ToggleKey } from "../../../../pages/main/model"
-import { useToggleState } from "../../../../shared/model/toggle-state.model"
+import { useToggleState } from "../../../../shared/model/ToggleStateContext"
 
 interface PostEditModalProps {
   post?: Post
