+++ src/shared/model/ToggleStateContext.tsx
@@ -1,5 +1,3 @@
-"use client"
-
 import { createContext, useContext } from "react"
 
 export type ToggleState = {
